% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/markdownit.R
\name{markdownit}
\alias{markdownit}
\title{Render markdown to HTML}
\usage{
markdownit(md, html = FALSE, xhtmlOut = FALSE, breaks = FALSE,
  langPrefix = "language-", linkify = FALSE, typographer = FALSE,
  quotes = "“”‘’", highlight = TRUE, inline = FALSE)
}
\arguments{
\item{md}{input markdown text}

\item{html}{enable HTML tags in source}

\item{xhtmlOut}{use \code{'/'} to close single tags (\code{<br />}).}

\item{breaks}{convert \code{'\\n'} in paragraphs into \code{<br>}}

\item{langPrefix}{CSS language prefix for fenced blocks}

\item{linkify}{autoconvert URL-like text to links}

\item{typographer}{enable some language-neutral replacement + quotes beautification}

\item{quotes}{double + single quotes replacement pairs, when typographer enabled and smartquotes on.
Set doubles to \code{'«»'} for Russian, \code{'„“'} for German.}

\item{highlight}{enable highlight.js syntax coloring}

\item{inline}{render inline (without paragraths & blocks)}
}
\description{
Renders markdown text to HTML using \href{https://github.com/markdown-it/markdown-it}{markdown-it}
}

